% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointr.r
\name{pointr}
\alias{pointr}
\title{Package 'pointr'}
\description{
The \pkg{pointr} package allows to work with pointers to R
  objects/selection in order to make the R code more readable and
  maintainable.
  The main function of the package are: \emph{\code{ptr()}} to
  create a pointer, \emph{\code{rm.ptr()}} to remove a pointer, and
  \emph{\code{where.ptr()}} to check the target object of a pointer.
}
